+++ src/main/java/com/example/team1_be/domain/Group/GroupService.java
@@ -10,7 +10,7 @@ import com.example.team1_be.domain.Group.DTO.Create;
 import com.example.team1_be.domain.Group.DTO.GetMembers;
 import com.example.team1_be.domain.Group.DTO.InvitationAccept;
 import com.example.team1_be.domain.Group.Invite.Invite;
-import com.example.team1_be.domain.Group.Invite.InviteService;
+import com.example.team1_be.domain.Group.Invite.Service.InviteService;
 import com.example.team1_be.domain.User.User;
 import com.example.team1_be.domain.User.UserService;
 import com.example.team1_be.utils.errors.exception.CustomException;
@@ -26,7 +26,6 @@ public class GroupService {
 
 	private final GroupRepository groupRepository;
 
-	@Transactional
 	public void create(User user, Create.Request request) {
 		if (!user.getIsAdmin()) {
 			throw new CustomException("매니저 계정만 그룹을 생성할 수 있습니다.", HttpStatus.FORBIDDEN);
@@ -40,14 +39,12 @@ public class GroupService {
 		userService.updateGroup(user, group);
 	}
 
-	@Transactional
 	public void invitationAccept(User user, InvitationAccept.Request request) {
 		if (user.getIsAdmin()) {
 			throw new CustomException("알바생 계정만 그룹에 가입할 수 있습니다.", HttpStatus.FORBIDDEN);
 		}
 
-		Invite invite = inviteService.findByCode(request.getInvitationKey());
-		inviteService.checkValidation(invite);
+		Invite invite = inviteService.findInvitation(request.getInvitationKey());
 
 		Group group = invite.getGroup();
 		userService.updateGroup(user, group);
@@ -60,7 +57,6 @@ public class GroupService {
 		return new GetMembers.Response(group, user, users);
 	}
 
-	@Transactional
 	public void creatGroup(Group group) {
 		groupRepository.save(group);
 	}
